requirements.txt
aiosqlite ~=0.18.0
ircrobots ~=0.6.1
pyyaml    ~=6.0

